sign-up api, move signInSocial, add key to  Feed Scroll
