+++ src/components/modules/SignIn/FormSignin/index.tsx
@@ -51,7 +51,6 @@ function FormSignin() {
       router.push('/');
     } catch (error) {
       console.log(error);
-      
     }
   };
