+++ src/components/modules/SignUp/Main/index.tsx
@@ -3,7 +3,7 @@
 import Title from 'antd/es/typography/Title';
 
 import FormSignUp from '@/components/modules/SignUp/FormSignUp';
-import SignInSocial from '@/components/modules/SignInSocial';
+import SignInSocial from '@/components/lib/SignInSocial';
 
 import * as S from './styles';
 function SignUp() {
